Added Boostrapping, fixed issues with missing values
Edited IMlibs to deal with the fact that some of the binding series
data didn’t have 8 points and was interfering with the conversion into
an array.  All directories for saving are pointing to my home directory
and will need to be changed.
Edited variantFun to do the bootstrapping (which is still slow), and
extract the data to a new table.  Right now 20% of the data has NaNs
for error bars, they are either varaitns that only have one cluster, or
variants that for some reason have NaN as a variant number.  I’m
investigating the second reason currently.  The filtering of variants
is just the value > 0.5, but should be changed depending on what is
working well for you.